+++ app/src/main/java/com/example/kaka/moviedb/CategoryAdapter.java
@@ -2,8 +2,6 @@ package com.example.kaka.moviedb;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 
 import com.example.kaka.moviedb.data.MovieReview;
@@ -23,7 +21,7 @@ public class CategoryAdapter extends FragmentPagerAdapter {
     private List<MovieReview> movieReviewList;
     private String overview;
 
-    public CategoryAdapter(Context context, FragmentManager fm, String overview,
+    public CategoryAdapter(Context context, android.support.v4.app.FragmentManager fm, String overview,
                            List<MovieReview> movieReviewList, List<MovieTrailer> movieTrailerList) {
         super(fm);
         mContext = context;
@@ -33,7 +31,7 @@ public class CategoryAdapter extends FragmentPagerAdapter {
     }
 
     @Override
-    public Fragment getItem(int position) {
+    public android.support.v4.app.Fragment getItem(int position) {
         if (position == 0) {
             Bundle bundle = new Bundle();
             bundle.putString(MovieOverviewsFragment.MOVIE_OVERVIEW_KEY, overview);
