+++ app/src/main/res/layout/movie_details.xml
@@ -1,127 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v4.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-    <LinearLayout
+    <FrameLayout
+        android:id="@+id/detail_container"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <android.support.constraint.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <android.support.v7.widget.CardView
-                android:id="@+id/cv_poster_image"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="10dp"
-                android:layout_marginTop="10dp"
-                android:elevation="@dimen/card_elevation"
-                app:cardCornerRadius="8dp"
-                app:cardUseCompatPadding="true"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintTop_toTopOf="parent">
-
-                <ImageView
-                    android:id="@+id/iv_movie_poster"
-                    android:layout_width="150dp"
-                    android:layout_height="200dp"
-                    android:contentDescription="@string/str_movie_poster"
-                    android:padding="@dimen/poster_padding"
-                    android:scaleType="fitXY"
-                    android:src="@drawable/not_found" />
-            </android.support.v7.widget.CardView>
-
-            <TextView
-                android:id="@+id/tv_movie_title"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="70dp"
-                android:layout_marginStart="10dp"
-                android:padding="@dimen/poster_padding"
-                android:textAppearance="?android:textAppearanceLarge"
-                android:textColor="@android:color/black"
-                android:textStyle="bold"
-                app:layout_constraintLeft_toRightOf="@+id/cv_poster_image"
-                app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toTopOf="@+id/cv_poster_image"
-                tools:text="Spider Man and Iron Man" />
-
-
-            <TextView
-                android:id="@+id/tv_release_date"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:drawablePadding="@dimen/poster_padding"
-                android:drawableStart="@drawable/ic_date_range_black_24dp"
-                android:gravity="start"
-                android:padding="@dimen/poster_padding"
-                android:textAppearance="?android:textAppearanceSmall"
-                android:textColor="@android:color/black"
-                app:layout_constraintLeft_toLeftOf="@+id/tv_movie_title"
-                app:layout_constraintTop_toBottomOf="@+id/tv_movie_title"
-                tools:text="March, 2017" />
-
-            <TextView
-                android:id="@+id/tv_rating"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:drawablePadding="@dimen/poster_padding"
-                android:drawableStart="@drawable/ic_star_black_24dp"
-                android:gravity="start"
-                android:padding="@dimen/poster_padding"
-                android:textAppearance="?android:textAppearanceSmall"
-                android:textColor="@android:color/black"
-                app:layout_constraintLeft_toLeftOf="@+id/tv_release_date"
-                app:layout_constraintTop_toBottomOf="@+id/tv_release_date"
-                tools:text="7.1" />
-
-            <TextView
-                android:id="@+id/tv_genre"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:padding="@dimen/poster_padding"
-                android:textAppearance="?android:textAppearanceSmall"
-                android:textColor="@android:color/black"
-                android:textStyle="italic"
-                app:layout_constraintLeft_toLeftOf="@+id/tv_rating"
-                app:layout_constraintTop_toBottomOf="@+id/tv_rating"
-                tools:text="Horror | Comedy | Drama" />
-
-        </android.support.constraint.ConstraintLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <android.support.design.widget.TabLayout
-                android:id="@+id/sliding_tabs"
-                style="@style/CategoryTab"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="10dp"
-                android:layout_marginStart="10dp"
-                android:layout_marginTop="10dp"
-                app:tabGravity="fill" />
-
-            <android.support.v4.view.ViewPager
-                android:id="@+id/viewPager"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="10dp"
-                android:layout_marginStart="10dp"
-                android:layout_marginTop="10dp" />
-
-        </LinearLayout>
-    </LinearLayout>
+        android:layout_height="match_parent" />
 </android.support.v4.widget.NestedScrollView>
